affine cipher: remove reference to an exit code
While it is idiomatic for shell scripts to indicate an error status
via exit code, doing so is uncommon in other languages. For generality,
the problem specifications should not instruct the test suites to
return an exit code.

This includes a bump to the patch version as the test suite has
not changed at all.