anagram: reimplement test case containing 'master'
Exercism has now renamed the main branch of every repo from `master` to
`main`, which makes it logically consistent to also remove references to
`master` in exercise tests. The only one was in the `anagram` exercise.

This commit adds a test case that tries to stay close to the one that it
reimplements, maintaining these properties:
- the `subject` is a six-letter word
- the `candidates` are three six-letter words
- the first and last `candidates` as anagrams
- the non-anagram has one letter in common with the `subject`

Closes: #1770