exercises/markdown: add extensive header parsing tests
As discussed in issue #1812, markdown exercise would benefit with more
extensive tests to ensure learners write dynamic code, capable of
handling all header depths there is.
To ensure their code is not too dynamic, I added the h7 level. It does
not exist in the HTML specification and therefore should be equal to the
input.

Resolves #1812.