grep: remove mentions of "output" and "single string"
The type implied by "single string" is not part of the essence of this
problem. Some tracks wish to ask for an ordered collection of lines,
rather than a single string containing the lines, and it seems that
there is no reason for the specification to forbid this.

The specific word "output" could also be misunderstood to mean "print on
standard output", though we're not yet currently aware of any track
tests that do indeed require this.

Closes https://github.com/exercism/problem-specifications/issues/1873