scale-generator: Require N+1 notes from N intervals
The problem has been explained in
https://github.com/exercism/problem-specifications/issues/1640:

All previous cases only required N notes from N intervals, but that
meant the last interval was completely ignored.
That causes confusion (what is the last interval there for?).

To fix this, all tests with the `interval` property have to be
reimplemented.

The `tonic` and `intervals` inputs are the same as the original.
The `expected` is the same as the original, except with an additional
last note indicated by the last interval, which was previously
ignored.

In conjunction with https://github.com/exercism/problem-specifications/pull/1884,
closes https://github.com/exercism/problem-specifications/issues/1640