sgf-parsing: Add individual tests of escaping/whitespace behaviour
11c36323-93fc-495d-bb23-c88ee5844b8c crams too much into one test case.
Split it up into multiple ones.

In addition, 11c36323-93fc-495d-bb23-c88ee5844b8c has behaviour that
violates the specification. The violation is that `\t` and `\n` (written
as `"\\t"` and `"\\n"` in the JSON string, respectively), do not hold
any sort of special significance in SGF, according to the specification:
https://www.red-bean.com/sgf/sgf4.html

Reimplement in 08e4b8ba-bb07-4431-a3d9-b1f4cdea6dab.

The reimplemented case is mostly as it was when the exercise was
originally implemented in
https://github.com/exercism/exercism/commit/7a5075bf96e2cafb29f2f883ff98b708be5594a6
and the reimplemented case is in accordance with the specification.

Note that the original case also got it wrong in that newlines should
remain newlines; this is corrected in
08e4b8ba-bb07-4431-a3d9-b1f4cdea6dab.