auto-format portions of some JSON files
this commit was made by having prettier format these files and
committing portions of the changes.

For now we want prettier to continue to ignore these files, but
eventually the plan would be to use prettier but to reformat specific
portions.

The portions in this commit are *not* the portions we would want to
reformat, so for these specific portions we are satisfied to accept
prettier's changes.