Delete bonus point sections from descriptions
The bonus points sections were a way of encouraging more discussion
and exploration in v1.

Discussion:
Now that we have closed mentoring rather than open discussions
it makes less sense to encourage discussion in this way.
We may find other ways of encouraging discussions, but then
it will be a product-level design choice rather than a textual
call to action from within the exercise description.

Exploration:
We tend to want people to explore the approach to solving the problem
rather than to explore how to expand the scope of the problem.
As such I think it's worth deleting the exploration aspect of the
bonus points sections.

Also, there are only a handful of exercises that had this sort of
section, so it's also neater in terms of consistency.