sgf-parsing: improve description flow
There is a bit of opinion on what makes better flow. The main opinion
of this commit:

High-level description of the goal should come earlier; that way the
high-level goal can be kept in mind when reading the details below.

So two pieces of text are moved:

First, the overall goal of the exercise.
Second, that all property values are Text.