Reformat one sentence per line (#2118)
This updates the exercise descriptions to have one sentence per line, to
conform to the Exerism Markdown Specification.

There are a couple of deprecated exercises that I didn't touch.
There are also a couple of files that have open PRs that are
fixing the one sentence per line thing as part of a bigger overhaul,
so I didn't touch those either.

There's a chance that I missed a spot here and there, as this task
was too messy to automate properly. (I could have done so, but then
it would have been roughly the same amount of work as doing it
manually, and it still probably would have gone wrong. So I didn't
bother with a script).