sgf-parsing: Replace incorrect info with correct info (#2114)
* sgf-parsing: improve description flow

There is a bit of opinion on what makes better flow. The main opinion
of this commit:

High-level description of the goal should come earlier; that way the
high-level goal can be kept in mind when reading the details below.

So two pieces of text are moved:

First, the overall goal of the exercise.
Second, that all property values are Text.

* sgf-parsing: Replace incorrect info with correct info

285d2e6da032805ac3ba7943ac9ef925afe07433 added outright incorrect info
("Incorrect" means it violates the specification found at
https://www.red-bean.com/sgf/sgf4.html#text).

It is replaced with correct info. As follows:

* Newlines are treated differently based on whether they come after a \
* SGF does **not** have escape sequences for whitespace characters, so
  it doesn't make sense to talk about escaped forms for them.
* There is no need or expectation for parentheses or semicolons to be
  escaped, so remove the sentence about that.

All statements in this commit are already being tested by tests added in
3c33c249121c16aa2a7fc3398a50349d8985f752, so no additional tests need be
added.

Co-authored-by: Katrina Owen <katrina.owen@gmail.com>