Clarify description of meetup
There has been a lot of confusion around this exercise, especially
for people who don't speak English as their native language.

This attempts to explain very clearly where 'teenth' comes from,
and how it pertains to recurring monthly meetup schedules.

In doing this I have also made an executive decision to _not_
change the structure of the inputs (see #1069).
This is because I don't believe that adding string parsing to
get the pieces of the general description makes this exercise
more interesting or fun.

In order to keep the exercise focused on just the date munging,
this instead changes the description to match more closely how
the inputs are given.

Closes #1786.
Closes #1069.