Reorder cases in roman-numerals canonical-data
This reorders the cases in the roman-numerals exercise.
The old order was sort-of increasing, but with some cases
that were out of order, clumped at the end.

This makes the order consistent, sorting numerically by the
input value.

This does not affect tracks that have implemented the exercise.
It's purely a nice-to-have for tracks that are adding the exercise
through the use of a generator.

I used jq to visually verify the order:

    cat exercises/roman-numerals/canonical-data.json \
    | jq -r .cases[].description

Order before:

    1 is I
    2 is II
    3 is III
    4 is IV
    5 is V
    6 is VI
    9 is IX
    27 is XXVII
    48 is XLVIII
    49 is XLIX
    59 is LIX
    93 is XCIII
    141 is CXLI
    163 is CLXIII
    402 is CDII
    575 is DLXXV
    911 is CMXI
    1024 is MXXIV
    3000 is MMM
    16 is XVI
    66 is LXVI
    166 is CLXVI
    666 is DCLXVI
    1666 is MDCLXVI
    3001 is MMMI
    3999 is MMMCMXCIX

Order after:

    1 is I
    2 is II
    3 is III
    4 is IV
    5 is V
    6 is VI
    9 is IX
    16 is XVI
    27 is XXVII
    48 is XLVIII
    49 is XLIX
    59 is LIX
    66 is LXVI
    93 is XCIII
    141 is CXLI
    163 is CLXIII
    166 is CLXVI
    402 is CDII
    575 is DLXXV
    666 is DCLXVI
    911 is CMXI
    1024 is MXXIV
    1666 is MDCLXVI
    3000 is MMM
    3001 is MMMI
    3999 is MMMCMXCIX