Rework two-fer documentation
This splits two-fer into separate introduction and instructions
documents.

It also rewords the exercise to feel a bit less stilted, and
(hopefully) a bit more friendly. It adds a bit of explanation
for the definition of the term two-fer and where it comes from,
as well as providing a bit of context ("story", if you will) for
the exercise itself.