Rework gigasecond documentation
This is an attempt at making the gigasecond
exercise easier to understand and a bit more
engaging.

The key change is to provide more context around
the idea of a gigasecond and why it is potentially
interesting.

It also tones down the mathiness in an attempt to
make it feel more casual, friendly, and approachable.

As part of this change we're splitting the
description.md file into two files:

- introduction.md which gives context, and
- instructions.md which tells the student what
  the task actually consists of.

This is a change that was documented as part of the v3
migration.

Co-authored-by: Jeremy Walker <jez.walker@gmail.com>