check_required_files: Only allow legacy or new formats
After an initial implementation in
https://github.com/exercism/problem-specifications/pull/2203,
ensuing discussion in
https://github.com/exercism/problem-specifications/pull/2211 tells us
that in fact we only want to allow two possibilities:

1. legacy format: Only description.md exists
2. new format: Only instructions.md and introduction.md exist.

Previously, no check for introduction was present, so one has to be
added.

If a an exercise has files that cause it to be a mix of the two formats,
that shouldn't be allowed either.