Fix formatting in two-fer instructions
I noticed that CI was unhappy in the javascript
track when we synced this exercise.

This reformats the table so that next time we sync
we don't undo the fix.

This is not a high-priority fix, so I won't be syncing
it to all the tracks; it can come in as part of the
next sync that brings in little fixes here and there.