Tweak wording in binary search introduction
The word number can refer to a musical number, like what someone would perform on stage,
or a mathematical number, like what you would use to count.

In this introduction we talk about songs, confusing the issue about what number refers to.

This adds a parenthetical to clarify which type of numbers we mean.