Tweak wording in binary search introduction (#2226)
* Tweak wording in binary search introduction

The word number can refer to a musical number, like what someone would perform on stage,
or a mathematical number, like what you would use to count.

In this introduction we talk about songs, confusing the issue about what number refers to.

This adds a parenthetical to clarify which type of numbers we mean.

* Update exercises/binary-search/introduction.md

* Add wikipedia links to binary-search

The numbers we chose as examples are interesting, but it's not obvious
why. The links help explain, to those who care to look.