Let multi line question example be one sentence (#2498)
Using multiple sentences is a bit ambiguous in this exercise, in
general. This is especially so, if we follow a sentence with a
question mark by a sentence with a dot.

To avoid that ambiguity gettting in the way of showcasing the
intended way to handle multi-line queries, let this example use just
one sentence, but let it spread over multiple lines.

  Disucssion: https://forum.exercism.org/t/bob-exercise-text-should-mention-linebreak-handling/13582/17