simple-cipher: complete description of the Vigenere
The description was only show the special case where the key is a
repeating letter, but not the general case. That was good for showing
how shift ciphers are a special case of substitution ciphers, but not
really for showing the full generality of the Vigenere.

Also, name the Vigenere cipher (which step 2 is all about): having the
name allows people to read more about it. Also, the Vigenere is a
special case of substitution ciphers, which are more general.